directory structure created
